Add a setup.py to the project
The `zine.py` script has been cleaned-up a bit, and refactored into
a set of functions, to be able to distribute it easily with as an
`entry_point`.
