+++ zine/static/style.css
